Final Version of the LED Function Sub-set
This version has no debugging print outs as it stuffed the text and the
output with a lot of extra Data.
LEDFunction in the other directory has every debugging line and one
alternative way of blinking for the LED.
For making exact time counting possible the debugging delay had to be
removed.